refactor(bundler): rename js bundler module to proper name
