+++ src/lib/bundler/js-bundler.js
